sonhos-lusiadas-backend/requirements.txt
# Backend Flask - Sonhos Lusíadas
# Dependências base do template
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# Dependências adicionais para análise literária
# Processamento de arquivos
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.9.0

# Processamento de linguagem natural tradicional
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0

# Análise de dados
pandas>=2.1.0
numpy>=1.24.0

# Visualização
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wordcloud>=1.9.2

# Utilitários
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Desenvolvimento
pytest>=7.4.0
